pyproject.toml
[project]
name = "scalarscope"
version = "0.1.0"
description = "ScalarScope - Evaluative Internalization Training Framework"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "mcp-tool-shop", email = "64996768+mcp-tool-shop@users.noreply.github.com" }
]
dependencies = [
    "numpy>=1.24",
    "pydantic>=2.0",
]

[project.optional-dependencies]
onnx = [
    "onnxruntime-gpu>=1.16",
    "transformers>=4.36",
    "sentencepiece>=0.1.99",
    "protobuf>=4.0",
]

torch = [
    "torch>=2.0",
]

dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

export = [
    "optimum[onnxruntime-gpu]>=1.16",
    "onnx>=1.15",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scalarscope"]

[tool.ruff]
line-length = 100
target-version = "py311"
